+++ js/picture.js
@@ -1,11 +1,58 @@
-import { renderComments, initCommentList } from './comment.js';
 import { isEscapeKey } from './util.js';
+
+const SHOWN_MAX_COMMENTS = 5;
 
 const bigPictureElement = document.querySelector('.big-picture');
 const bodyElement = document.querySelector('body');
 const closeBigPicture = bigPictureElement.querySelector('.big-picture__cancel');
 
+const commenstListElement = bigPictureElement.querySelector('.social__comments');
+const commentsTotalElement = bigPictureElement.querySelector('.social__comment-total-count');
+const commentCountElement = bigPictureElement.querySelector('.social__comment-shown-count');
+const commentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
+const commentElement = document.querySelector('#comment').content.querySelector('.social__comment');
+
+let commentsCountShown = 0;
+let comments = [];
+
+const createComment = ({avatar, message, name}) => {
+  const newComment = commentElement.cloneNode(true);
+
+  newComment.querySelector('.social__picture').src = avatar;
+  newComment.querySelector('.social__picture').alt = name;
+  newComment.querySelector('.social__text').textContent = message;
+
+  return newComment;
+};
+
+
+const renderComments = () => {
+  commentsCountShown += SHOWN_MAX_COMMENTS;
+
+  if (commentsCountShown > comments.length) {
+    commentsLoaderElement.classList.add('hidden');
+    commentsCountShown = comments.length;
+  } else {
+    commentsLoaderElement.classList.remove('hidden');
+  }
+
+  const fragment = document.createDocumentFragment();
+  for (let i = 0; i < commentsCountShown; i++) {
+    const comment = createComment(comments[i]);
+    fragment.append(comment);
+  }
+  commenstListElement.innerHTML = '';
+  commenstListElement.append(fragment);
+
+  commentCountElement.textContent = commentsCountShown;
+  commentsTotalElement.textContent = comments.length;
+
+};
+
+const onCommentsLoaderClick = () => renderComments();
+
 const hidePicture = () => {
+  commentsCountShown = 0;
   bigPictureElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
   document.removeEventListener('keydown', onDocumentKeydown);
@@ -34,12 +81,17 @@ const showPicture = (pictureData) => {
   bigPictureElement.classList.remove('hidden');
   bodyElement.classList.add('modal-open');
   document.addEventListener('keydown', onDocumentKeydown);
-  renderComments(pictureData.comments);
-  initCommentList();
+
+  comments = pictureData.comments;
+  if (comments.length > 0) {
+    renderComments();
+  }
 
   renderPicture(pictureData);
 };
 
 closeBigPicture.addEventListener('click', onClosePictureButton);
+commentsLoaderElement.addEventListener('click', onCommentsLoaderClick);
 
 export { showPicture };
+
